Denest NonRecs in SpecConstr for more specialisation (#22277)
See Note [Denesting non-recursive let bindings].

Fixes #22277. It is also related to #14951 and #14844 in that it
fixes a very specific case of looking through a non-recursive let binding in
SpecConstr.

If applied to master pre-loopification changes (see !9141), there is virtually
no impact on NoFib or ghc/alloc perf, but it fixes a bad interaction between
Loopification and SpecConstr.