Refactor getCaretDiagnostic
In order to properly create this error message, we need access to the
line of code that caused the error. We need it without the carets
though, so we factor the part of `getCaretDiagnostic` out that grabs
the code

====
Add a new value ctor for `NotInScopeError`

Types.hs:
We need a new value ctor for `NotInScopeError` so we can tell if the
error is specifically related to associated types.

Codes.hs:
Because we made a new ctor, we also need a new error code for it.

====
Create the `UnknownAssociatedType` error

Env.hs:
The only way we can tell that we're dealing with an associated type is
via the `what` parameter. We change the parameter type to a
string, deferring SDoc creation to within `lookupInstDeclBndr` so
we can examine what `what` is. If it's `associated type`, we throw out
the `err` contained in `mb_name` and create the more specific
`UnknownAssociatedType` error. Otherwise we just do the normal thing.

Bind.hs: Pass our `what` param in as a plain string since
`lookupInstDeclBndr` requires that now

====
Construct the associated type error message

Now that we have everything in place, we can create our error message
for unknown associated types.

====
Accomodate for the `Maybe String`

I'm not exactly sure why the srcSpan would fail to generate a code,
but if it does, we use a simplified error message

====
Reword error and use `GhcHint`

I might've gone too far with this one, but we can just switch it back
if so. I thought of new wording for the error that might be a little
more direct.
We also utilize GhcHint for the suggestion on how to resolve the error.

Hint.hs:
Create a new value ctor for GhcHint: `SuggestDeclareAssociatedType`
which instructs the user how to resolve `UnknownAssociatedType` error.

Types.hs:
We add `RdrName` as a parameter to `UnknownAssociatedType` so we can
hand it off to our new GhcHint type

Env.hs:
We pass the RdrName to UnknownAssociatedType so we have it available
when creating the hint in `scopeErrorHints`

Errors/Ppr.hs:
Reword the error message and remove the resolution hint since it's been
moved to GhcHint
In `scopeErrorHints`, we pass the params from `UnknownAssociatedType`
to `SuggestDeclareAssociatedType`.

Hint/Ppr.hs:
We add in our SDoc wording for the `SuggestDeclareAssociatedType` hint

====
Update some failing test cases for the new wording

====
Add comments and use explicit fields

Use explicit fields On `UnknownAssociatedType` and
`SuggestDeclareAssociatedType` ctors, just so the constructors are a
little better documented