DmdAnal: Make exnDiv behave like topDiv
... except for the purposes of dead code eliminaiton.

As far as DmdAnal is concerned, `topDiv` and `exnDiv` used to differ simply in
their `defaultArgDmd`. NoFib suggests that nothing is lost by making the
`defaultArgDmd` of `exnDiv` return `topDmd` instead of `absDmd`.
(Of course the main purpose of `exnDiv` remains: dead code elimination.)

In the interest of simplification we should make `topDiv` and `exnDiv` as
similar as possible.